refactor!: move STM+related defs to reaction
