+++ tests/channels/test_jpsi_to_gamma_pi0_pi0.py
@@ -6,10 +6,10 @@ import logging
 
 import pytest
 
+from expertsystem.reaction import InteractionTypes, StateTransitionManager
 from expertsystem.reaction.combinatorics import (
     _create_edge_id_particle_mapping,
 )
-from expertsystem.ui import InteractionTypes, StateTransitionManager
 
 logging.basicConfig(level=logging.ERROR)
 logging.getLogger().setLevel(logging.ERROR)
