+++ tests/channels/test_y_to_d0_d0bar_pi0_pi0.py
@@ -9,7 +9,7 @@ import pytest
 
 from expertsystem.amplitude.canonical_decay import CanonicalAmplitudeGenerator
 from expertsystem.amplitude.helicity_decay import HelicityAmplitudeGenerator
-from expertsystem.ui import (
+from expertsystem.reaction import (
     InteractionTypes,
     StateDefinition,
     StateTransitionManager,
