+++ tests/conftest.py
@@ -1,9 +1,9 @@
 import pytest
 
-from expertsystem import ui
 from expertsystem.particle import ParticleCollection
+from expertsystem.reaction import load_default_particles
 
 
 @pytest.fixture(scope="module")
 def particle_database() -> ParticleCollection:
-    return ui.load_default_particles()
+    return load_default_particles()
